+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - [#9](https://github.com/FUSAKLA/promruval/pull/9) Upgraded to Go 1.17
 
 ### Added
- - [#10](https://github.com/FUSAKLA/promruval/pull/9) New validator `validateAnnotationTemplates` for more info see [the docs](docs/validations.md#validateannotationtemplates)
+ - [#10](https://github.com/FUSAKLA/promruval/pull/10) New validator `validateAnnotationTemplates` for more info see [the docs](docs/validations.md#validateannotationtemplates)
+ - [#11](https://github.com/FUSAKLA/promruval/pull/11) New validator `forIsNotLongerThan` for more info see [the docs](docs/validations.md#forisnotlongerthan)
+
 
 ## [v1.3.2] - 2020-12-09
 ### Fixed